generate core Game/Pot/Plant models
the goal here is to have a Game object that holds the state of a current
game (how much money the player has, how many days have passed, what
pots exist), and then delegate responsibility to the different objects
as needed. Right now money is straightforward but perhaps that will be
part of a catalog or budget catalog later on down the line.

Pots have a position so we know where to render them and will hold both
a plant and its current needs. If a pot is watered or in sunlight then
then its plant will be able to grow.

Plants will mostly be a state machine of "can I grow?". I've gone with
adding a `type` field for STI in order to subclass the defined behavior
of indivudal species of plants. I could have gone with a catalog but
then I would have write the lookup code for it, and adding a new class
isn't much more effort than, say, a configuration file with the same
data.