add some plant types!
these sub classes exist only to define the different parameters for each
plant. Again, this could have potentially been a separate table for
archetype lookups but I think it's an easier cogitive abstraction (even
if we have to deal STI) to think of a plant as a "Calathea" or a
"Philodendron". Plus if later on there is some complexity to add (say,
for propoation) then this is an easy way to define separate behaviors
for different plant types.