initialize and render games
lays out a basic game "board" and defines a way to create a new game
based on it. Currently a game is pretty much a series of pots on a grid,
but it will expand to track the size of the store, positions of other
other items, and maybe a catalog.

Already there's an opportunity to refactor this code by introducing a
Board object to abstract the responsibility of rendering the game, but
it's good to not jump into the wrong abstraction off the bat.