determine sunlight level for pots based on position
I've made a really limited definition of sunlight to start off with only
a few values and based soley on the y position of a pot. We can imagine
that our store has big windows at the very bottom (something that
graphics can relay later) and so pots that are closest to the bottom of
the board get a high amount of sunlight while pots closest to the top of
the board get a low amount.

Later I can imagine variable influencing these (additional rooms with
different light levels across the whole room for example) or more
complexities (direct vs indirect) but as long as we have *a* model for
sunlight we can simulate growth.